fixup MDEV-29181: protect from concurrent DROP FOREIGN KEY
We can have some redundant elements in fk_ref_list, which could be dropped
concurrently, before referenced tables are prelocked.

We have to update the list after prelocking, to make sure they are removed.

Note that we can't rely on the order of the list filled by
get_parent_foreign_key_list, because internal referenced_set is a dynamic
hash, and therefore it could be rehashed, if too many elements were
removed.