fixup MDEV-29181 Assertion foreign_table->get_ref_count() != 0 failed
Change the fts initialization to more granular. It should be done only for
tables with cascade/set null foreign keys, so that we won't do it for
tables with ref_count==0. That was previous fix of some other assertion.

But even that wasn't enough, since sql chooses tables to open more
accurately, so the operation should match the foreign key type, too.

In sql layer, thd->trg_event_map will tell us, what operations are possible
by the command. In innodb, it will be enough to check upd_node->is_delete.