MDEV-29639 Regression only
Problem:
========
If parallelism is enabled on a replica, Seconds_Behind_Master can
spike high in cases of delayed or infrequent transactions. This is
because a parallel slave updates last_master_timestamp at the end of
an event, rather than the beginning, to make for a less confusing
value of Seconds_Behind_Master during times of high concurrency.
However, when dealing with delayed or infrequent transactions,
Seconds_Behind_Master will use the last committed transaction on the
slave in its calculation, leading to potentially very large values.